restructure and nodenext
